Creates Scores class and High Score popup
This creates the Scores class which includes a method to check if a new
score entry fits into the top scores list. It also includes a getScores
method that can be used to display the scores in the TopScoresActivity
(needs to be implemented). This also creates the popup window for the
user to enter in their information. Still need to include logic for when
this window pops up.