Incorporates High Score popup into game logic
Ideas for how to pass the text from the popup to the Score class have
been incorporated into comments but they are not working properly yet.
The EditText widget of the popup is giving me problems - cannot figure
out how to extract the text from the widget and return it as a string
without crashing the program.