Creates customized buttons
The splash screen and start game buttons have been reconfigured to be
backgroundless and use a custom style, with the same font used
elsewhere in the app. An XML file helps change the color of the button
text when touched, to give the user some feedback that they are in fact
clicking them.